WIP: explore svg elements with createSVGElement
This PR is a sample PR introducing the concept of a call to
createElementNS (via createSVGElement). This allows SVG elements to be
rendered in DLight.

Current limitations:
- only works with `svg` and `circle` elements

Future questions:
- would it be possible to add all SVG element tag names to a
defaultSVGTags set? what about tags that exist in both namespaces (e.g.
`a`, `title`, `script`, `font` etc.)
- perhaps detecting if a tag is a child of an SVG element tag should
automatically cause DLight to tag it as belonging to the SVG namespace?